examples/projects/azure-aci-deploy/requirements.txt
boto3
deltalake
pandas
azure-identity
azure-mgmt-containerinstance
